Move tar.go to testutils
There are various CodeQL findings from inside tar.go related to
extracting potentially malicious tar files. However tar.go was only ever
used inside test code. Hardening test code is futile as any malicious
user can modify the code base of their fork in any way they want and
thus attack the CI system.

Moving tar.go to testutils and putting in a comment stating that the
code must not be used in production makes it clear that this code is
only to be used for testing. The CodeQL findings can then be marked as
"wontfix".
